refactor(#394): extract shared logout hook
