+++ apps/frontend/src/features/auth/hooks/index.ts
@@ -1,2 +1,3 @@
 export { useFormState } from './useFormState';
 export { useAuthBootstrap } from './useAuthBootstrap';
+export { useLogout } from './useLogout';
